#21 refactor: Rs 네이밍에 Read 제거
