+++ motimo-api/src/main/java/kr/co/api/goal/docs/GoalControllerSwagger.java
@@ -2,9 +2,9 @@ package kr.co.api.goal.docs;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import kr.co.api.common.rs.CreateIdRs;
 import kr.co.api.goal.rqrs.GoalCreateRq;
-import kr.co.api.goal.rqrs.GoalListReadRs;
+import kr.co.api.goal.rqrs.GoalIdRs;
+import kr.co.api.goal.rqrs.GoalListRs;
 import kr.co.api.goal.rqrs.GoalUpdateRq;
 
 import java.util.UUID;
@@ -12,11 +12,11 @@ import java.util.UUID;
 @Tag(name = "목표 API", description = "목표 관련 API 목록입니다")
 public interface GoalControllerSwagger {
     @Operation(summary = "목표 생성 API", description = "목표를 생성합니다.")
-    CreateIdRs create(GoalCreateRq rq);
+    GoalIdRs create(GoalCreateRq rq);
 
     @Operation(summary = "목표 수정 API", description = "목표를 수정합니다.")
     void update(String id, GoalUpdateRq rq);
 
     @Operation(summary = "목표 목록 API", description = "목표 목록을 조회합니다.")
-    GoalListReadRs readList(UUID userId);
+    GoalListRs readList(UUID userId);
 }
