+++ motimo-api/src/main/java/kr/co/api/goal/rqrs/GoalListRs.java
@@ -4,8 +4,8 @@ import io.swagger.v3.oas.annotations.media.Schema;
 
 import java.util.List;
 
-public record GoalListReadRs(
+public record GoalListRs(
     @Schema(description = "목표 목록")
-    List<GoalItemReadRs> goals
+    List<GoalItemRs> goals
 ) {
 }
