+++ motimo-api/src/main/java/kr/co/api/point/docs/PointControllerSwagger.java
@@ -1,10 +1,8 @@
 package kr.co.api.point.docs;
 
 import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import kr.co.api.point.rqrs.PointReadRs;
+import kr.co.api.point.rqrs.PointRs;
 import kr.co.api.security.annotation.AuthUser;
 
 import java.util.UUID;
@@ -13,5 +11,5 @@ import java.util.UUID;
 public interface PointControllerSwagger {
 
     @Operation(summary = "현재 획득한 포인트 조회 API", description = "사용자가 획득한 포인트를 조회합니다.")
-    PointReadRs read(@AuthUser UUID userId);
+    PointRs read(@AuthUser UUID userId);
 }
