+++ motimo-api/src/main/java/kr/co/api/point/rqrs/PointRs.java
@@ -2,7 +2,7 @@ package kr.co.api.point.rqrs;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 
-public record PointReadRs(
+public record PointRs(
     @Schema(description = "사용자가 현재 획득한 포인트", example = "1000")
     int point
 ) {
